app/api/v1/sentences/requirements.txt
fastapi==0.99.0
mangum==0.11.0
pydantic==1.10.9
python-dotenv==1.0.0
python-multipart==0.0.6
starlette==0.27.0
typing-extensions==4.9.0
google-cloud-texttospeech==2.14.1
google-generativeai==0.8.5
boto3==1.34.34
requests==2.31.0
uvicorn
